Finalizing basic mithril testing
