+++ basic/index.html
@@ -1,7 +1,8 @@
 <!DOCTYPE html>
 <html>
 	<head>
-		<meta charset="utf-8">
+        <meta http-equiv="X-UA-Compatible" content="IE=edge">
+        <meta charset="utf-8">
 		<title>Mithril Comment</title>
         <link rel="stylesheet" type="text/css" href="../css/style.css">
     </head>
@@ -13,8 +14,8 @@
 
     </div>
 
-    <script type="text/javascript" src="./json_shimmed.js"> </script>
-    <script type="text/javascript" src="../node_modules/es-5-shim/es-5-shim.js"> </script>
+    <!--<script type="text/javascript" src="./json_shimmed.js"> </script>-->
+    <!--<script type="text/javascript" src="../node_modules/es-5-shim/es-5-shim.js"> </script>-->
     <script type="text/javascript" src="../node_modules/mithril/mithril.js"> </script>
     <script type="text/javascript" src="./app.js"> </script>
