+++ browserify/index.html
@@ -1,6 +1,7 @@
 <!DOCTYPE html>
 <html>
 <head>
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta charset="utf-8">
     <title>Mithril Comment</title>
     <link rel="stylesheet" type="text/css" href="../css/style.css">
