+++ browserify/wiki.js
@@ -5,16 +5,16 @@ var wiki = {};
 
 wiki.data =
 
-    wiki.model = function(){
-        this.title = m.prop(wiki.data().title);
-        this.content = m.prop(wiki.data().content);
-        this.version = m.prop(wiki.data().version)
+    wiki.model = function(content){
+        this.title = m.prop(content.title);
+        this.content = m.prop(content.content);
+        this.version = m.prop(content.version)
     }
 // Long way of binding the data to view
 wiki.controller = function(){
     var self = this;
-    this.data = m.prop({});
-    m.request({method: "GET", url: './wiki.json'}).then(self.data);
+    //this.data = m.prop();
+    this.data = m.request({method: "GET", url: './wiki.json', type : wiki.model});//.then(self.data);
 
     this.edit = m.prop(false);
 
@@ -22,8 +22,8 @@ wiki.controller = function(){
     this.toggleView = function(){
         if(self.edit()){
             // save
-            self.data().version++;
-            logs.List().push(new logs.singleLog("wiki", self.data().version));
+            self.data().version(self.data().version()+1);
+            logs.List().push(new logs.singleLog("wiki", self.data().version()));
             self.edit(false);
         } else {
 
@@ -41,17 +41,17 @@ wiki.view = function (controller) {
             m(".panel-heading", [
                 m(".row", [
                     m(".col-md-9", [
-                        m("span", "Change Title: "), m("input.form-control", { onchange: m.withAttr("value", function(store){ controller.data().title = store; }), value: controller.data().title} )
+                        m("span", "Change Title: "), m("input.form-control.editWikiTitle", { onchange: m.withAttr("value", controller.data().title), value: controller.data().title()} )
                     ]),
                     m(".col-md-3.cm-wikiBar", [
                         m(".btn-group", [
-                            m("button.btn.btn-sm.btn-default[type='button']",{ onclick : controller.toggleView },  m("i.fa.fa-save", " Save"))
+                            m("button.saveWikiButton.btn.btn-sm.btn-default[type='button']",{ onclick : controller.toggleView },  m("i.fa.fa-save", " Save"))
                         ])
                     ])
                 ])
             ]),
             m(".panel-body", [
-                m("textarea.ht-wiki-edit", { onchange: m.withAttr("value", function(store){ controller.data().content = store; }), value: controller.data().content} )
+                m("textarea.ht-wiki-edit", { onchange: m.withAttr("value", controller.data().content), value: controller.data().content()} )
             ])
         ])
     } else {
@@ -59,19 +59,19 @@ wiki.view = function (controller) {
             m(".panel-heading", [
                 m(".row", [
                     m(".col-md-9", [
-                        m("h2.panel-title", controller.data().title)
+                        m("h2.panel-title", controller.data().title())
                     ]),
                     m(".col-md-3.cm-wikiBar", [
                         m(".btn-group", [
-                            m("button.btn.btn-sm.btn-default[type='button']",{ onclick : controller.toggleView }, m("i.fa.fa-pencil", " Edit"))
+                            m("button.editWikiButton.btn.btn-sm.btn-default[type='button']",{ onclick : controller.toggleView }, m("i.fa.fa-pencil", " Edit"))
                         ])
                     ])
                 ])
             ]),
             m(".panel-body", [
-                m("p#wiki-preview", controller.data().content)
+                m("p#wiki-preview", controller.data().content())
             ]),
-            m(".panel-footer", " Version " + controller.data().version)
+            m(".panel-footer.wikiVersion", " Version " + controller.data().version())
         ])
     }
